Move font code into a separate FontManager class
